requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------

matplotlib>=3.2.2
numpy==1.22.3
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1


tqdm>=4.41.0
tensorflow==2.8.0
pycocotools==2.0.4
slidingwindow==0.0.14



# Fair MOT
yacs
cython-bbox
progress
motmetrics
lap
openpyxl
tensorboardX
fvcore


# PCPA
wget
scikit-learn